Use `encodeURIComponent` for username URLs
This partly addresses #4604, but not for usernames that end in '?'. For other usernames that just have characters that need to be URI-encoded to be part of a valid URL (like "Wikster808%"), this makes the Students tab and User Profile page largely usable. The one problem I've noticed that isn't fixed is that you can't navigate directly to the student articles view for such a username — eg, http://localhost:3000/courses/test/test_long_course_(test)/students/articles/Wikster808% — because the `history` util throws an error when calling `createBrowserHistory()`.