More fixes for student usernames that break the Student Details view
This gets even closer to fully fixing https://github.com/WikiEducationFoundation/WikiEduDashboard/issues/4604 . The only remaining issues with the example usernames I've looked at are that you can't navigate directly to a URL for a user with a %  in the actual username (eg, Wikster808%). I think the percent issue is a bug with the `history` component used by react-router version 5, and it might be fixed with react-router@6 / history@5.

The solution here isn't very elegant, as ideally we want to encode/decode using the same algorithm that the router is using (and without re-implementing it).

There might also be remaining characters that are possible in usernames but aren't handled yet. In particular, some of the unhandled reserve characters noted here — https://github.com/remix-run/history/issues/461#issuecomment-295825802 — are valid in MediaWiki usernames (+ = & $). Those might require extending the selectUserByUsernameParam() function.