Improve error message for non-existing user
If a user account exists on some wikis but not on the home wiki for a course, trying to add that user by their username will result in an error. Including the wiki domain in the error message should make it more understandable when this happens, and make the solution more obvious.