Handle another format of API response for MediaWiki spam blocking
The previous update for the new API response for edits blocked due to matching the spam list didn't cover the format used when there's only a single match, leading to some remaining unhandled errors.

This should unblock on-wiki course page mirroring for all the rest of the courses that were getting stopped as spam (approximately 2 of them, based on browsing error logs).