Partially fix week numbering in Milestones component
This adds an offset to the numbering so that the first week of the Timeline is numbered the the same way as shown in the Milestones component, for courses with time between course start and timeline start.

It doesn't handle vacation weeks, though, so the numbering can still be off from the Timeline versus the Milestones component on the Overview tab, if there are weeks with no meetings in the middle of the timeline.

Fixing the remaining problem will require a bit of a refactor, probably by moving a canonical representation of the Timeline — including week titles and dates — to redux with a selector. Currently, the logic of inserting empty weeks based in the middle of the timeline lives in `timeline.jsx`.