Update more gems
I think this is all the gems that we can compatibly update at the moment, except for I18n-js which needs to be updated in concert with the JS package. Some gems are held back because they are dependencies for the deprecated `paperclip` gem, which we need to migrate off of.