Fix ArticleViewer's handling of articles with "&" in the title
ArticleViewer would not render an article initially, instead showing a 404 error message that (incorrectly) indicates a WhoColor failure. Once authorship data is received, it would then render correctly with authorship highlighting.

The '&' character gets encoded as '%26' with `encodeURIComponent` but gets left alone (since it's part of the ?foo=a&bar=2 parameter syntax) with `encodeURI`. So, to correctly construct a parse URL for a Wikipedia article with a title includes '&' (among other control characters, probably), we need to use `encodeURIComponent` on the title rather than `encodeURI on the whole URL. (We can't use both, or it will double-encode the control characters in the title.)