Improve URL generation for articles ending in ?
Both the article links (from the Students tab, for example) and the draft and bibliography links were broken for article titles ending in a question mark (because MediaWiki can't tell that it's part of the title rather than the beginning of URL params). So, unlike most other characters, the ? at the end of an article needs to be converted to `%3F` to form a valid link to the article.

We already handled this for usernames, using CGI.escape. The same strategy should work smoothly for article titles. The downside is that the URLs are not as clean and readable as they could be, since CGI.escape does other conversions as well that are not necessary (like commas).