Extend peer review training with more explicit guidance
This should help address the common problem of students not being able to find their classmates' work to review. Since in most cases, the problem is that the classmate has not posted a draft or has posted it somewhere else, having some explicit guidance for that situation will hopefully cut down on confusion.