Fix bug with adding a tracked wiki when wikidata is home wiki
If Wikidata (or another wiki with `null` as the language) is the home wiki, then the normalizeWikis function results in a duplicate copy of the home wiki, because one copy has `null` for the language and the other has `www`. This results in a 422 error when you attempt to save the course after adding a tracked wiki.

Ensuring that 'www' is added for the language in place of `null` makes the normalizeWikis utility function work as intended.