Add workaround spec for UploadImporter error handling path
I believe it's still possible for errors like this to happen in production, but the one example I had that drove the original implementation has been changed on Wikimedia Commons (because the file name also caused problems on-wiki), so I can't actually replicate it for the spec anymore. I haven't been able to find any other live examples to test against.

This spec is really just a workaround to make sure that code path gets run in the test suite, although it does somewhat approximate the flow of when this error would actually occur.