Generate an Alert when a WP Med article is assigned without proper training
Goal: To generate an Alert when a WikiProject medicine article is assigned
for a course that doesn't have the medical content training module.

How to find such WP med article ? Their talk page belongs to the
Category:All_WikiProject_Medicine_article category.
We query wp for that. After grabing the last assignments, we check
for med articles without the training.
For each case -> an alert is raised.

- add a new sub type of alert
- add to log in constant update
- add a lib monitor file
- the matching spec
- the matching preview
- add a factory to create a training module
- fix two unexpected spec bugs(training modules created but non deleted in a spec)
  that cause the temp creation of a training module to fail because of
  duplicate slug name in the med alert spec.