Add support for a :no_max_users flag
This lets us (manually from the console) set a :no_max_users flag on a course, which lets that course bypass the 150 user limit for mass enrollment.

The limit is intended to prevent courses that track too many super-active users, but in many cases there are large-user-count events that won't cause problems (depending on the wiki and the activity level of the most active participants), so this lets us work around the limit on request, if it seems safe.

The initial use case for this is GLAM Macedonia events like https://outreachdashboard.wmflabs.org/courses/GLAM_Macedonia/WIKI_CLUBS_2021_(2021)