Add basic skeleton of CheckSandboxes service
This is a barebones first pass at a system for periodically checking the content of student sandbox pages to flag possible problems. It uses the REGEXes from https://en.wikipedia.org/wiki/User:Headbomb/unreliable.js to identify possibly-unreliable sources.

This iteration is just meant to be run manually via the code in the analytics script to generate a CSV, but once we try it out for recent courses and see what the data looks like, it might make sense to build this into an update cycle and create Alerts for these violations.