Fix two bugs related to Wikidata assignments.
The change to AssignButton fixes a confusing warning when you try to assign an item in a Wikidata course (`warning_untracked_wiki`) because this component treated the assigned article as www.wikidata.org but the home wiki language as `null`.

The change to article_finder_utils fixes the problem of Wikidata search queries being blocked by CORS policy. I'm not sure why this hasn't been an issue in the past and why Wikipedia queries (still) work without specifying `jsonp` but Wikidata queries don't. There might have been a change in configuration on Wikidata recently.