Fix display of recent blog posts after theme update
We've changed the WordPress theme on wikiedu.org, and this changed the format of the RSS feed that we use for blog posts, adding the 'screen-reader-text' span as part of the 'Read More' link. (The link also used to be called 'continued', and that link was not in a new paragraph tag.) This tweak hides the screen-reader-text span, and also makes the images a little taller. Not perfect, but looks good enough.