Fix for rendering Rails parts on a wrong tabs
Most of the Campaign tabs are a combination of Rails views and React components. When switching between Rails-React only React tabs, Rails parts of a previous page were staying on a new tab. That is solved by searching for  IDs of those parts and removing their content (if found) on ComponentDidMount.