Updated regex of relative path
Before this, a relative path was defined as a path not having the word
http/https in it. However, some requests, like the delete assignments
one, pass in a query parameter with a URL, which contains the word
https. This causes the relative path check to fail, even though
/assignments is clearly a relative path. Following this change, a path
will be called relative only if it doesn't START with http/https.