Don't pass `null` to the markdown renderer in Milestones component
I'm not sure how, but sometimes a Milestone block with just a training module and no content has occurred in production. I can't replicate it, because blanking the text (or never entering any) for a Block typically results in an empty string rather than null, which doesn't cause a problem. But this should at least make sure that everything still renders if we end up with that situation again.

Resolves #4718