Make sure ArticleImporter updates titles for existing articles
This fixes the problem where an article already exists in the Dashboard database under a different title but the same mw_page_id (typically, because the page got moved since the Dashboard stopped updating for whatever course edited it before), so adding an Assignment for that article would not connect the assignment to an Article record. This meant that the Available Articles or Assigned Articles listing would indicate 'this article does not exist', which was confusing for users. (This would fix itself once a student edited the article, because then the edited article would get its title updated by one of the routine update processes.

Resolves #4607