Add frontend support for Wikidata 'References added' stat
This also adds a guard for undefined stats, so that we can easily roll out new stats as part of WikidataOverviewStats (etc) and have it work smoothly for courses that don't have an up-to-date CourseStat record that includes a newly added stat. This way, 'References added' will simply not show up for a given course until it's gone through an update that adds the count for that value (even if it's zero).