Fix 'Request Account' feature
This was accidentally broken during the conversion from jquery to fetch, with the code that dug into the response to find the object within the users array getting lost along the way. We didn't catch it because `request_account_spec.rb` doesn't pass on CI anyway, and I didn't check it locally.