Use Browser Router instead of Router
Since React Router version 6, location(that is the one given by the
useLocation hook) must be used inside of a Router. Since the navbar is
not inside a router, it can't use location. Instead, it now uses the
global window.location. Because of this change, the manual
implementation of history is no longer required, and has been deleted