Fix assignment tag replacement bug
This was deleting the newlines between the section header and the tag, when updating an existing tag. For example: https://en.wikipedia.org/w/index.php?title=Talk:Hair_analysis&diff=next&oldid=1074176778&diffmode=source

That's because the preceding newlines (but not the header itself) were part of the match being substituted for the case of updating an existing template.