Fix rendering of "Download stats" button on Campaign overview
For a Campaign with more than the minimum details in the Details section, and/or content in the Program Template section, the server-rendered elements were covering up the "Download stats" button. This change puts the button above all the details, immediately below the stats. Not ideal, perhaps, but it's a quick fix for the problem.