Show recently-created campaigns rather than all active ones on Explore page
The current list of campaigns and their stats is not very useful, and when it's not cached it can take dangerously long to load. Limiting it to just recently-created ones, and just 10 of them, should make it both more interesting for surfacing interesting new things and also alleviate performance issues.