Skip loading or rendering articles list for huge campaigns
The Articles tab of the default Miscellanea campaign on Programs & Events Dashboard has terrible performance, sometimes taking a minute or two to load (and causing poor performance for other requests in the meantime). This is a quick solution to the problem: just don't load any articles if there are too many.