Make sure CSV builder doesn't error when Article record is missing
In production, we've been hitting errors when the system tries to create a campaign CSV. It appears to be caused by missing Article records. (I'm not sure why there are course revisions that don't have an Article record.)

This just lets it handle that situation gracefully by skipping the CSV lines for missing articles.

Also, there was no reason to use `map` instead of `each` for the course revisions loop.