Fix tests
The browser won't let us read from the coverage data while there's an
alert which hasn't been accepted. This makes sure all alerts are being
accepted.

Even after doing this, one of tests failed. I found that this was
because it was trying to parse empty text as JSON. I must've missed this
when I converted AJAX calls to fetch. This test didn't fail earlier
because there was an alert on the screen which was never accepted(and so
the response was never even parsed)