only highlight key which is currently used for sorting
To check which key is used for sorting, we check if it has an
`order` property. When the user clicks on the header text or selects the
key from the dropdown, `order` is added. But the `order` property is not
removed from the previously selected keys. This means multiple key headers
are highlighted. This fixes that