Add some quick workarounds for a flaky test
I think recent flapping of the course_cloning_spec, which come right after this one in the test suite, are due to JS errors that don't get retrieved during this test and so cause the following one to fail.

Locally, I get occasional failures of this test run in isolation because of an apparent race condition in the fetching of special users, but there may be a similar one at the end that is causing CI failures. The one at the beginning is theoretically an actual bug, although practically it's unlikely for a user to click 'Approve Course' before the initial fetches are completed. The one at the end is probably not a realistic bug outside of test conditions.