Rework JS error output from rspec feature tests
We've been getting JS errors that are causing flapping tests, but the full error message doesn't display on CI and it's been hard to pin down the cause. Maybe relying on the `expect` to print the error message is part of the problem.

Instead, we'll print all the JS warnings/errors before the `expect` and see if we can get more details.