Tweaks to Uploads view to prevent JS errors
There were a couple possible JS errors, mostly benign, that were causing tests to fail intermittently. In different test runs, it returns different data for usage and categories, and when categories are present, it would cause a missing key error. Also (undetected by the tests), in List mode it could throw an error in List view when the UploadViewer modal was activated, because that modal (a div not a tr) would be part of a table.