rewrite the range function as a generator
Since `courseWeeks` can be negative, a loop from 0 to `courseWeeks` is not enough. This is why we need to use the `range` function.
This makes it possible to iterate over the range function in a for loop without creating an array of all the numbers in the range.
Also removes the need for the arrow function.