Update PlagiabotImporter to use new toolforge tool
The eranbot API has been down since 'Stretch' jobs were disabled on toolforge. Rather than attempt to migrate `eranbot` to a newer webservice engine — without breaking any of the cron jobs that are still running on that tool — I built a simple Sinatra webservice to access the database of suspected plagiarism and serve the recent suspected diffs and ithenticate report urls that the Dashboard needs.