Improve search.json response time
This fixes an N+1 query problem with the course search results. On dashboard.wikiedu.org, a search for 'fall' took about 18 seconds and returned about 2500 results. With this change, it's about 6 seconds and one query (~300ms). There's still something very slow about rendering this, though, which scales with the number of results.

Removing the unused 'description' field reduces the amount of json data sent (from 2.7mb to 1.2mb in the 'fall' example), although that's not a significant contribution to response time.