fix tests
To make the tests pass, I've made it so that the toDate function emulates some of momentjs' questionable choices
like using undefined and null to indicate current time. MomentJS also converted undefined and null to 0
in the add and subtract functions. date-fns returns an invalid date in such cases, which caused another
test to fail. This is a workaround for that.