fix timeline length
Earlier, the timeline length was being calculated by moment(end - start).weeks()
This would give incorrect results because end - start returns an integer - the number of
milliseconds between the two dates(which can then be converted to weeks).
Calling moment on it converts it to a date instead of a duration.