Return UTC date string instead of date
While doing some more reading about dates in Javascript, I found that the Date constructor always returns dates in
the local timezone. So having a date object in a different timezone is not really possible. Since moment-js had its
own wrapper Date object, it didn't have this problem. date-fns operates on the native Date object, so it limited in that
regard.

So while we can't get a Date object with UTC timezone, getting a string with UTC timezone is possible. This is what toISOString()
does