fix tests
A couple of tests were failing because we were accessing the incorrect attribute, while in another one we weren't setting the date.
There's still a JS error for the user profile page but that will require a bit more work and refactoring to solve satisfactorily.
For now, this suppresses that error.