Add target="_blanks" to typical links in ArticleViewer
This cleanly handles standard wikilinks in Wikipedia articles, so that clicking one will open the Wikipedia article in a new tab instead of navigating the user away from the Dashboard.

Wikidata links, however, are formatted differently, with a `title` attribute before the `href` and relative links instead of absolute ones... so this change doesn't affect Wikidata items, which have broken links in ArticleViewer anyway.