Modify relativeLinkMatcher to work for Wikidata links
On Wikidata, all the internal links include a `title` attribute before the `href`, so ArticleViewer was not converting them into absolute links (and so, they would point to nonexistent Dashboard pages).

This tweak adds matching for an optional `title="something"` attribute before the `href`, making the link fixing (and the `target="_blank"` addition from ParsedArticle.jsx) work for Wikidata.