Quick workaround for non-functional links in /unsubmitted_courses
Only the first link in the table of unsubmitted courses actually works. I'm not sure how long this has been broken or exactly why — it seems likely to be relatedc to relying on the `data-link` attribute and how rails JS helpers work with that — but in any case, we'll eventually rewrite this page in React. Meanwhile, this makes it usable.