Fix formatting bug with date-fns
This was throwing an error because toDate from `date-fns` does not accept strings, so it would break rendering of any course page where the 'Last Reviewed' flag includes a timestamp (ie, any course being viewed by an admin who is listed as staff for that course and the course has been marked as reviewed). Per the warning emitted just before the error, using parseISO fixes it.