Fix deprecation warning for Enumerable.sum
CampaignCsvBuilder#sum_wiki_columns was throwing warnings whenever it would sum a column that included non-numeric values, since Rails has deprecated the Rails implementation of Enumerable.sum in favor of the native Ruby version.

Using 0 rather than an empty string as the default fallback value of a stat_row in CourseWikidataCsvBuilder fixes this problem.

In production, there are no CourseStat records with an empty string as a value, so this should be a safe change even after the behavior of .sum changes in Rails 7.1.