Add workaround for MediaWiki query bug
I've hit this bug before because I reported it, but it just turned up in production again, causing a course to make an infinite loop of queries during its update process (and therefore, never complete an update). This workaround should be a good-enough fix for it, even if it never gets fixed on the MediaWiki side.