Fix some JS errors in tickets_spec
We weren't actually using the hidden selected target user input for the GetHelpButton form, as we pull the selected user from state anyway.

Also, I couldn't trigger it except in the test, but the ticket spec would throw an error from the selector at the end, right after deleting the ticket.