Fix bug where 'dismiss' icon isn't usable in error notification
A user reported that, while using Article Finder, undismissable error notifications would gradually take over the screen. The errors were from ORES API request timeouts, but the URL that threw the error was so long that it pushed the 'dismiss' icon wasn't visible. Setting 'overflow: overlay' adds a horizontal scroll bar instead when there is a string that is too long but won't wrap within an error message, so the dismiss icon should remain usable.