Minor adjustments to Event Center API
Per discussion with WMF engineers, I've removed the not-strictly-necessary `verify_organizers` step for the `update_event_participants` endpoint. I've also changed the `organizer_usernames` parameter; it is now plural, and requires an array of usernames rather than a single username. This will make it compatible with planned changes on the Event Center end of things to support multiple organizers.