Fix calculation of words and references for namespace-limited stats
The course top-line stats for words and references added weren't respecting the namespace-specific stat tracking, because CoursesUsers records were not accounting for tracked versus untracked namespaces. With this change, only articles in tracked namespaces will contribute to the user's stats in the CoursesUsers record, and therefore in the top-line course stats.