Minimal working version of bibliography reviews
I don't like passing down `dispatch`, so I guess I should implement `connect` in `ButtonNavigation`.

All the logic of the alert itself — which article it relates to, and what message to display to link to it — should be moved to the server. This can probably be done with `subject_id` to connect it to the Assignment record.