Fix counting of sandbox and draft revisions in CoursesUsers
The recent change to implement tracked namespaces resulted in basing CoursesUsers calculations only on tracked namespaces, which meant that the 'user' and 'draft' character sums went to zero (as those namespaces are not tracked by default).

This restores the previous behavior for revision count, along with userspace and draft character sums, while keeping the new behavior of including all tracked namespaces in the main "mainspace" character counts.

This makes the column name "character_sum_ms" a little misleading, as it can now include non-mainspace namespaces if they are tracked. However, I think that's a minor issue; in spirit, we treat all tracked namespaces as main.