remove unnecessary event listener from the Articles Tab
These event listeners were being used to close the modal when the user clicked outside of it. However, instead of using just one event listener for the currently open modal,
there was a listener for each modal. This was causing several redundant events to be fired even when there was no modal open.

This attempts to streamline the process by using a single event listener for the active modal, and removes it when the modal is closed.