Add sandbox existence checks for Assignments
These are the backend methods we can use to check which expected sandboxes related to an Assignment have already been created.

It also refactors the setting of default sandbox URLs so that it's possible to maintain separate sandboxes for two Assignments to the same article in the same class, without the `before_save` callback overwriting it to the shared sandbox.