Update some screenshots in training modules
Wikipedia is about to switch to the Vector 2022 skin as the default for new users, so we're updating our training screenshots to use Vector 2022 wherever relevant. Most screenshots were very similar, so we uploaded new ones over the old ones on Commons. The two changed were more substantially different, so we uploaded them as a separate file.