eagerly load users and their survey notifications
Earlier, these were being fetched on demand which resulted in a lot of unnecessary queries. This is now fixed. The way this works is by loading the user data in app/views/surveys/_question_group.html.haml and then using the prefetched data in the various survey methods. I've had to add a fallback to the old way of fetching the data since the prefetched data is not available in all cases - just while viewing the results.