Fix bug with loading tickets with no sender record
This came up in production today, specifically the message.sender.real_name throwing an error, but just to be safe, I added safe navigation for all the attributes where a similar situation might occur. Some of these may not be necessary, but this is easier than trying to exhaustively rule out the possibility.