Use 4-byte character sanitization in ProtectedArticleMonitor
For the last week, this part of the ConstantUpdate job has been failing because of a collation error when the array of titles is used in a database query, presumably because there's a protected article that has characters that we haven't tested for in this particular code path. Using the same forced conversion we use for this purpose elsewhere should get it working again.