Try some things to avoid 422 error when cloning courses
We've recently encounted to instances in production of a bug where attempting to clone a course results in a 422 error. This is being triggered because of a supposed invalid wiki. I think it might have to do with an un-persisted duplicate courses_wikis record for the same wiki, although I haven't been able to replicate the bug in a development environment.

The two changes here should basically result in the same intended behavior as before, but with a few added hedges against un-saved CoursesWikis records.