Add sandbox statuses to assignments.json
This isn't used by the frontend yet, but it will make the sandbox status available for incorporating into the frontend soon.

(I also noticed that :article_title was part of the .call all the beginning, but then was being overwritten by the version that replaces underscores. Removing it from the .call just changes the output by moving the position of article_title property within a single assignment's JSON object.)