Add handling for non-existing Bibliography sandboxes, fix bug
This is the first step of front-end integratoin with sandbox status checking. It makes the Bibliography link red if the page doesn't exist yet, and removes the preload if it does exist.

Also, I caught a couple of bugs with the sandbox assignment checking.

First bug: it was looking for bibliographies at /bibliography rather than the actual location we link to, /Bibligraphy (with the capital B).

Second bug: the mediawiki API didn't like a pagename that included a URL-encoded character (because of parentheses in the page name). I added a URI component decoding step to work around that.