Add redlink behavior to peer review links
This extends the redlink styling to the peer review links, and adds status updates for the corresponding bibliography and draft sandboxes for Assignment records in the REVIEWING role, so that the links for reviewers will also behave correctly in terms of redlinks versus normal links.

Also, it adds mouseover text to explain the meaning of a redlink here.