Add back non-enzyme tests that were deleted by mistake
With the upgrade to React 18, we're deleting all the component tests, which were very rarely helpful. However, the rest of the tests that don't actively use enzyme are mostly functional unit tests and some of them, especially the utils tests, are pretty useful for preventing regressions and documenting the intended uses and edge cases for specific things.