Fix for bug with assignments that have draft data but no status updates
The intended behavior of starting from the first sandbox status didn't work for an assignment where the sandbox status had been updated, resulting in a status of `null` on the frontend and disabling the ability to use the progress tracker to change the status. Hopefully this restores the intended behavior.