Add option to disable sandbox links and progress tracker via flags
We are doing a pilot for having courses skip using sandboxes in Fall 2023. We will manually add the :no_sandboxes flag for these courses, which will remove the sandbox links from MyArticles and the Students tab, and also remove the ProgressTracker and related UI, since that is built around the sandbox workflow.

I included some refactoring for using course properties to control the presence of ProgressTracker UI without just checking course type over and over in the JS, which makes the logic a little clearer.