Switch to LiftWing API for importing articlequality data
The ORES API is being deprecated, and the same data from the articlequality models now comes from LiftWing.

The data format is largely the same, but LiftWing does not support batch requests, this first implementation simply makes a request for each revision serially. For performance, we will likely need to implement concurrent requests, but we'll also need to work with the hourly request limits (50k by default, more if an authenticated client with proper permissions is used). The results from individual revisions are combined to match the format of batch results from ORES, to minimize the changes necessary in RevisionScoreImporter. I've also worked around the different format of results for missing/TextDeleted revisions, so that the error results can also be combined into the batch structure.

Resolves #5465