Add another guardrail for very long update courses
If we discover that a course in production is so resource-intensive to update that it will cause the worker to die before successfully finishing an update, then adding the `very_long_update` flag to the course record is not sufficient to ensure that it doesn't end up clogging the system; when a worker dies, the incomplete job gets put back in the front of the queue, so when the worker restarts it will just try it again.

With this change, adding the flag will also cause an update to return immediately when the worker takes the job from the queue, so we can more easily deal with cases where a course is clogging the system.