Fix namespace limitation on Category records
Whoops! When we set the default to namespace 0 for Category queries, this broke the system for categories that are on the corresponding Talk pages of articles, such as those derived from WikiProject templates.

The NoMedTrainingForCourseAlert's have not been going out for the last 4 months since this namespace default was set.