Try a workaround for 'schedule course updates' jobs getting rejected
There seems to be a problem with the lock in P&E Dashboard production, such that the cron for ScheduleCourseUpdatesWorker isn't actually getting the job into the queue, even though there is no existing lock I can find. (The changelog says 'Limit exceeded'.) I'm hoping that this tweak to the sidekiq-unique-jobs option will get it working again.