Restore previous lock option
None of the tweaks to the options worked to unblock the scheduling of this worker, but I manually deleted the uniquejobs keys in redis that were causing the problem in production. I'm not sure why they were not showing up on the 'locks' list in the sidekiq UI or why deleting all locks via the UI didn't do the trick. My best guess is that the offending key(s) were from an earlier version of sidekiq-unique-jobs which weren't recognized by the current version but were still enforcing lock behavior.