Add support for meta.wikimedia.org
Some users on P&E Dashboard want to track projects that focus on edits to meta. This adds 'meta' to the list of available languages, updates the Replica interface to use the right DB name, and updates the frontend so that meta.wikimedia.org is the only 'meta' entry that appears in the wiki selects.