Add support for Wikidata default "mul" labels
The wikidata_course_spec broke recently because it tests for fetching and using Wikidata entity labels, using Q42 as the example. Wikidata recently rolled out the special 'mul' language. We can use 'mul' as an additional fallback (actually, a better fallback than 'en'), but we'll continue also to use 'en', as usage of the new 'mul' slot will not be immediate.