Add troubleshooting documentation about database encoding issues
I just changed the dashboard.wikiedu.org production database to use UTF8mb4 for the ticket_dispenser_messages table. Emails that included emoji were triggering database encoding errors and failing to create tickets.

For good measure, I also added an emoji to the EmailProcessor specs. It passes either way, but it fails the same way as in production if the table is converted to UTF8mb3.