Update real name on student's courses via /onboarding
We sometimes get requests from instructors to add or update a student's real name, for cases where the student didn't initially add a real name or didn't use their actual name. While we don't want to let an instructor access a real name simply by adding a username to their course, if the student takes action to update their real name, that's a good enough signal that they intend to share it with the instructors of their courses. So with this change, getting the correct real name on a course page becomes self-service for the student, via /onboarding.