Simple initial implementation of exercise sandbox verification
This just adds a server-side check for the existing of the expected Exercise Sandbox page before allowing an exercise to be marked complete, with an error notification explaining the problem.

At some point, we might want to make #verify_exercise_sandbox more sophisticated, with a minimum size or checking for specific hallmarks of a given exercise having been completed correctly, but this might be good enough.