Handle Category index collision for PagePile cats
The PagePileApi tries to reconcile cases where PagePile returns data from a different wiki than what the Category record expects, but updating the Category to whatever wiki PagePile says that pile is from. However, this can result in an index collision if a category for that pile already exists with the correct wiki.

This fix basically extends the remediation strategy. An alternative might be to ensure that pilepile ID is unique at the time the category is added, perhaps using a similar "get_or_create_by" strategy.